fix to allow more than 8 axis

+++ FluidNC/src/Machine/Axes.h
@@ -74,7 +74,7 @@ namespace Machine {
 
         static void set_disable(int axis, bool disable);
         static void set_disable(bool disable);
-        static void step(uint8_t step_mask, uint8_t dir_mask);
+        static void step(uint16_t step_mask, uint16_t dir_mask);
         static void unstep();
         static void config_motors();
